+++ packages/react/src/components/tooltip/Tooltip.module.css
@@ -1,4 +1,36 @@
 @import '~hds-core/lib/helsinki.css';
+
+.iconTooltip {
+  width: 1.4375em;
+  height: 1.4375em;
+}
+
+.iconClose {
+  width: 1.125em;
+  height: 1.125em;
+}
+
+.buttonTooltip {
+  position: absolute;
+  top: 0;
+  right: 0;
+  line-height: 0;
+  fill: var(--hds-ui-color-black-80);
+}
+
+.buttonClose {
+  position: absolute;
+  top: 0;
+  right: 0;
+  padding: 1.25em;
+  width: ;
+  fill: var(--hds-theme-color-primary);
+}
+
+.alternative .buttonClose {
+  fill: var(--hds-theme-color-secondary);
+}
+
 .tooltip {
   margin: 0.75em 0 1.25em;
   border-top: 0.625em solid var(--hds-theme-color-primary);
@@ -24,7 +56,7 @@
   vertical-align: middle;
 }
 
-.bodyText {
+.tooltipBody {
   composes: text-sm from '~hds-core/lib/helsinki.css';
   padding: 0.75em 0;
 }
