+++ packages/react/src/index.ts
@@ -5,3 +5,4 @@ export { default as Section, SectionProps } from './components/section/Section';
 export { default as Koros, KorosProps } from './components/koros/Koros';
 export { default as ImageWithCard, ImageWithCardProps } from './components/imagewithcard/ImageWithCard';
 export { default as Columns, ColumnsProps } from './components/columns/Columns';
+export { default as Tooltip, TooltipProps } from './components/tooltip/Tooltip';
